Ported most of HBMAME to base MAME, except for the following:
Neo-Geo (due to using old slot implementation, though the ROM definitions are preserved for later usage)
Monaco GP (non-trivial modifications to core, and not a real simulation in the first place)
A few drivers removed from base MAME due to license incompatibility

This repository will now be home to my future additions, and the old one is now inactive but preserved.
I will first compile using this repo when 0.192 is released.